+++ src/sentineltray/__init__.py
@@ -1,6 +1,6 @@
 """SentinelTray package."""
 
 __author__ = "Christian Martin dos Santos"
-__version__ = "5.0.1rc1"
-__version_label__ = "5.0.1 rc.1"
-__release_date__ = "17-04-2026"
+__version__ = "5.1.0"
+__version_label__ = "5.1.0"
+__release_date__ = "18-04-2026"